# app/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a payment gateway module with Swagger documentation.",
        "title": "Payment Gateway Module API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/v1/payments": {
            "post": {
                "description": "Creates a new payment with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment details",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment created successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Payment with this reference already exists",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/payments/{id}": {
            "get": {
                "description": "Retrieves payment details by payment ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment found",
                        "schema": {
                            "$ref": "#/definitions/domain.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID format",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Payment": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "reference"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "ETB",
                        "USD"
                    ]
                },
                "id": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "reference"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "ETB",
                        "USD"
                    ]
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusSuccess",
                "StatusFailed"
            ]
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "in: body",
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer"
                        },
                        "description": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        },
                        "params": {}
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}